+++ nylas/client/restful_models.py
@@ -591,6 +591,7 @@ class Event(NylasAPIObject):
         "original_start_time",
         "object",
         "message_id",
+        "ical_uid",
     ]
     datetime_attrs = {"original_start_at": "original_start_time"}
     collection_name = "events"