+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased (dev)
 ----------------
 * Add `ical_uid` to Event model, only available on API version 2.1 or above. See https://headwayapp.co/nylas-changelog/icaluid-support-132816
 * Add RoomResource model, available at `/resources` API endpoint
+* Add free/busy information, available at `/calendars/free-busy` API endpoint
 
 v4.9.0
 ------